# dev/transit-back/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Documentation de l'API Transit",
        "description": "Documentation de l'API trans.it générée avec Swagger",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local API Server"
        }
    ],
    "paths": {
        "/api/admin/stats": {
            "get": {
                "tags": [
                    "Admin"
                ],
                "summary": "Récupérer les statistiques globales de la plateforme",
                "operationId": "3781a63877fb9197d56d7ecff645c3a5",
                "responses": {
                    "200": {
                        "description": "Statistiques récupérées avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "totalUsers": {
                                            "type": "integer",
                                            "example": 150
                                        },
                                        "totalTransfers": {
                                            "type": "integer",
                                            "example": 340
                                        },
                                        "totalDownloads": {
                                            "type": "integer",
                                            "example": 1205
                                        },
                                        "totalStorage": {
                                            "type": "number",
                                            "format": "float",
                                            "example": 500.25
                                        },
                                        "activeTransfers": {
                                            "type": "integer",
                                            "example": 45
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/history/{id}": {
            "get": {
                "tags": [
                    "Historique"
                ],
                "summary": "Récupérer un historique par son ID",
                "operationId": "2a471b446bbba31e9cf89cb16f1011c6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de l'historique",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Détails de l'historique",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "ID": {
                                            "type": "integer"
                                        },
                                        "Id_Lien": {
                                            "type": "integer"
                                        },
                                        "Action": {
                                            "type": "string"
                                        },
                                        "Date_Action": {
                                            "type": "string",
                                            "format": "date-time"
                                        },
                                        "Adresse_IP": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Historique introuvable"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/links/{id}/history": {
            "get": {
                "tags": [
                    "Historique"
                ],
                "summary": "Récupérer l'historique d'un lien spécifique",
                "operationId": "0090dca2b7fa97395a991b3594d9d982",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID du lien",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste des historiques pour ce lien",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "ID": {
                                                "type": "integer"
                                            },
                                            "Id_Lien": {
                                                "type": "integer"
                                            },
                                            "Action": {
                                                "type": "string"
                                            },
                                            "Date_Action": {
                                                "type": "string",
                                                "format": "date-time"
                                            },
                                            "Adresse_IP": {
                                                "type": "string"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Aucun historique trouvé pour ce lien"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/history": {
            "post": {
                "tags": [
                    "Historique"
                ],
                "summary": "Ajouter une entrée dans l'historique",
                "operationId": "633888a2a862582946606944201f6a65",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "Id_Lien",
                                    "Action",
                                    "Adresse_IP"
                                ],
                                "properties": {
                                    "Id_Lien": {
                                        "type": "integer",
                                        "example": 12
                                    },
                                    "Action": {
                                        "type": "string",
                                        "example": "download"
                                    },
                                    "Adresse_IP": {
                                        "type": "string",
                                        "example": "192.168.1.1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Historique créé avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erreur lors de la création"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
...
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "apiKey",
                "description": "Enter token in format: Bearer {token}",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
    "security": [
        {
            "bearerAuth": []
        }
    ]
}